// src/data/api/datatyper/personalia.tsx
export enum Gradering {
    UKJENT = 'UKJENT',
    UGRADERT = 'UGRADERT',
    FORTROLIG = 'FORTROLIG',
    STRENGT_FORTROLIG = 'STRENGT_FORTROLIG',
    STRENGT_FORTROLIG_UTLAND = 'STRENGT_FORTROLIG_UTLAND'
}

export enum RelasjonsBosted {
    SAMME_BOSTED = 'SAMME_BOSTED',
    ANNET_BOSTED = 'ANNET_BOSTED',
    UKJENT_BOSTED = 'UKJENT_BOSTED'
}